requirements.txt
# Unified Requirements File
# Consolidated from requirements-core.txt, requirements-automation.txt, and requirements-ai.txt
# Python 3.11 compatible

# Core Web Framework - Updated to working versions
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.5

# Database
psycopg2-binary==2.9.9
psycopg[binary]==3.1.12
sqlalchemy==2.0.42

# Authentication & Security
python-jose==3.3.0
python-jose[cryptography]
pyjwt==2.10.1
bcrypt==4.3.0
passlib==1.7.4

# Data Validation - Updated to working versions
pydantic==2.11.7
email-validator==2.1.1

# Network Automation (Netmiko, pyATS/Genie) - Updated versions
netmiko==4.6.0
pyats==25.6
genie==25.6

# Configuration & Environment - Updated version
python-dotenv==1.1.1

# Scheduling
python-crontab==3.3.0

# HTTP & WebSockets - Updated to working versions
aiohttp==3.12.15
httpx==0.28.1
httpcore==1.0.9
async-timeout
websockets==15.0.1

# Utilities
itsdangerous

# AI/LLM Dependencies (Optional - install separately if needed)
# Note: These may have conflicting dependencies with the core application
# openai>=1.7.1,<2.0.0
# groq==0.6.0
# crewai==0.5.0
# langchain==0.1.0
# langgraph==0.0.18
# openrouter>=0.1.0
# chromadb>=0.4.0
# anthropic
